editor/libeditor/text/tests/test_bug318065.html

<!DOCTYPE HTML>
<!-- ***** BEGIN LICENSE BLOCK *****
   - Version: MPL 1.1/GPL 2.0/LGPL 2.1
   -
   - The contents of this file are subject to the Mozilla Public License Version
   - 1.1 (the "License"); you may not use this file except in compliance with
   - the License. You may obtain a copy of the License at
   - http://www.mozilla.org/MPL/
   -
   - Software distributed under the License is distributed on an "AS IS" basis,
   - WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License
   - for the specific language governing rights and limitations under the
   - License.
   -
   - The Original Code is Plaintext Editor Test code
   -
   - The Initial Developer of the Original Code is
   - Graeme McCutcheon <graememcc_firefox@graeme-online.co.uk>.
   - Portions created by the Initial Developer are Copyright (C) 2011
   - the Initial Developer. All Rights Reserved.
   -
   - Contributor(s):
   -
   -
   - Alternatively, the contents of this file may be used under the terms of
   - either the GNU General Public License Version 2 or later (the "GPL"), or
   - the GNU Lesser General Public License Version 2.1 or later (the "LGPL"),
   - in which case the provisions of the GPL or the LGPL are applicable instead
   - of those above. If you wish to allow use of your version of this file only
   - under the terms of either the GPL or the LGPL, and not to allow others to
   - use your version of this file under the terms of the MPL, indicate your
   - decision by deleting the provisions above and replace them with the notice
   - and other provisions required by the GPL or the LGPL. If you do not delete
   - the provisions above, a recipient may use your version of this file under
   - the terms of any one of the MPL, the GPL or the LGPL.
   -
   - ***** END LICENSE BLOCK ***** -->
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=318065
-->

<head>
  <title>Test for Bug 318065</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>  
</head>

<body>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=318065">Mozilla Bug 318065</a>
  <p id="display"></p>
  <div id="content" style="display: none">
  </div>

  <pre id="test">
    <script type="application/javascript">

      /** Test for Bug 318065 **/
      SimpleTest.waitForExplicitFinish();
      SimpleTest.waitForFocus(function() {
        var expectedValues = ["A", "", "A", "", "A", "", "A"];
        var messages = ["Initial text inserted",
                        "Initial text deleted",
                        "Undo of deletion",
                        "Redo of deletion",
                        "Initial text typed",
                        "Undo of typing",
                        "Redo of typing"];
        var step = 0;

        function onInput() {
          is(this.value, expectedValues[step], messages[step]);
          step++;
          if (step == expectedValues.length) {
            this.removeEventListener("input", onInput, false); 
            SimpleTest.finish();
          }
        }

        var input = document.getElementById("t1");
        input.addEventListener("input", onInput, false);
        var input2 = document.getElementById("t2");
        input2.addEventListener("input", onInput, false);

        input.focus();

        // Tests 0 + 1: Input letter and delete it again
        synthesizeKey("A", {});
        synthesizeKey("VK_BACK_SPACE", {});

        // Test 2: Undo deletion. Value of input should be "A"
        synthesizeKey("Z", {accelKey: true});

        // Test 3: Redo deletion. Value of input should be ""
        synthesizeKey("Z", {accelKey: true, shiftKey: true});

        input2.focus();

        // Test 4: Input letter
        synthesizeKey("A", {});

        // Test 5: Undo typing. Value of input should be ""
        synthesizeKey("Z", {accelKey: true});

        // Test 6: Redo typing. Value of input should be "A"
        synthesizeKey("Z", {accelKey: true, shiftKey: true});
      });
   </script>
  </pre>

  <input type="text" value="" id="t1" />
  <input type="text" value="" id="t2" />
</body>
</html>
